Build assets-registry for general consumption
