+++ packages/assets/package.json
@@ -11,7 +11,14 @@
   "homepage": "https://github.com/facebook/react-native/tree/HEAD/packages/assets#readme",
   "keywords": ["assets", "registry", "react-native", "support"],
   "bugs": "https://github.com/facebook/react-native/issues",
-  "engines": {
-    "node": ">=18"
-  }
+  "exports": {
+    "./package.json": "./package.json",
+    "./path-support": "./dist/path-support.js",
+    "./registry": "./dist/registry.js"
+  },
+  "files": [
+    "dist",
+    "path-support.js",
+    "registry.js"
+  ]
 }
