+++ packages/assets/registry.js
@@ -10,29 +10,4 @@
 
 'use strict';
 
-export type PackagerAsset = {
-  +__packager_asset: boolean,
-  +fileSystemLocation: string,
-  +httpServerLocation: string,
-  +width: ?number,
-  +height: ?number,
-  +scales: Array<number>,
-  +hash: string,
-  +name: string,
-  +type: string,
-  ...
-};
-
-const assets: Array<PackagerAsset> = [];
-
-function registerAsset(asset: PackagerAsset): number {
-  // `push` returns new array length, so the first asset will
-  // get id 1 (not 0) to make the value truthy
-  return assets.push(asset);
-}
-
-function getAssetByID(assetId: number): PackagerAsset {
-  return assets[assetId - 1];
-}
-
-module.exports = {registerAsset, getAssetByID};
+module.exports = require('./dist/registry');
